refactor: update imports to use locale module and add locale normalization tests
